+++ ceammc/CHANGELOG.md
@@ -9,6 +9,7 @@
   - an.tempo~: tempo analizer based on Aubio library
 - new properties:
   - lfo.+saw~, lfo.saw~, lfo.+tri~, lfo.tri~, lfo.+square~, lfo.square~, lfo.+pulse~, lfo.pulse~: @phase and @pause properties added
+  - ui.knob: @show_value property added
 - new methods:
   - seq.phasor: [phase( method added to set current phase
   - lfo.+saw~, lfo.saw~, lfo.+tri~, lfo.tri~, lfo.+square~, lfo.square~, lfo.+pulse~, lfo.pulse~: [reset( method added to reset to initial phase value (@phase)